Move addon enablement to addons module + lint fixes
